Move `vanilla_dicom_image` fixture to `conftest`
For easier reuse
